+++ src/components/PeopleFilters/PeopleFilters.tsx
@@ -1,10 +1,17 @@
-import cn from 'classnames';
+import React, { useEffect, useMemo } from 'react';
 import { useSearchParams } from 'react-router-dom';
 
-import { SearchLink } from '../SearchLink';
 import { Person } from '../../types';
-import { useEffect, useMemo } from 'react';
 import { SearchParams, getSearchWith } from '../../utils/searchHelper';
+import {
+  filterBySex,
+  filterByQuery,
+  filterByCenturies,
+} from '../../utils/filterHelpers';
+import { SexFilter } from './SexFilter';
+import { QueryFilter } from './QueryFilter';
+import { CenturyFilter } from './CenturyFilter';
+import { SearchLink } from '../SearchLink';
 
 type Props = {
   people: Person[];
@@ -12,32 +19,6 @@ type Props = {
 };
 
 const centuryArr = ['16', '17', '18', '19', '20'];
-
-const filterBySex = (people: Person[], sex: string) => {
-  return sex ? people.filter(person => person.sex === sex) : people;
-};
-
-const filterByQuery = (people: Person[], query: string) => {
-  const normalizedQuery = query.trim().toLowerCase();
-
-  if (!query) {
-    return people;
-  }
-
-  return people.filter(person =>
-    person.name.toLowerCase().includes(normalizedQuery),
-  );
-};
-
-const filterByCenturies = (people: Person[], centuries: string[]) => {
-  if (!centuries.length) {
-    return people;
-  }
-
-  return people.filter(person =>
-    centuries.some(century => Math.ceil(person.born / 100) === +century),
-  );
-};
 
 export const PeopleFilters: React.FC<Props> = ({ people, onFilter }) => {
   const [searchParams, setSearchParams] = useSearchParams();
@@ -52,12 +33,10 @@ export const PeopleFilters: React.FC<Props> = ({ people, onFilter }) => {
     () => filterBySex(people, sex),
     [people, sex],
   );
-
   const filteredPeopleByQuery = useMemo(
     () => filterByQuery(filteredPeopleBySex, query),
     [filteredPeopleBySex, query],
   );
-
   const filteredPeopleByCenturies = useMemo(
     () => filterByCenturies(filteredPeopleByQuery, centuries),
     [filteredPeopleByQuery, centuries],
@@ -93,76 +72,14 @@ export const PeopleFilters: React.FC<Props> = ({ people, onFilter }) => {
   return (
     <nav className="panel">
       <p className="panel-heading">Filters</p>
-
-      <p className="panel-tabs" data-cy="SexFilter">
-        <SearchLink
-          className={cn({ 'is-active': !sex })}
-          params={{ sex: null }}
-        >
-          All
-        </SearchLink>
-        <SearchLink
-          className={cn({ 'is-active': sex === 'm' })}
-          params={{ sex: 'm' }}
-        >
-          Male
-        </SearchLink>
-        <SearchLink
-          className={cn({ 'is-active': sex === 'f' })}
-          params={{ sex: 'f' }}
-        >
-          Female
-        </SearchLink>
-      </p>
-
-      <div className="panel-block">
-        <p className="control has-icons-left">
-          <input
-            data-cy="NameFilter"
-            type="search"
-            className="input"
-            placeholder="Search"
-            value={query}
-            onChange={handleQueryChange}
-          />
-
-          <span className="icon is-left">
-            <i className="fas fa-search" aria-hidden="true" />
-          </span>
-        </p>
-      </div>
-
-      <div className="panel-block">
-        <div className="level is-flex-grow-1 is-mobile" data-cy="CenturyFilter">
-          <div className="level-left">
-            {centuryArr.map(century => (
-              <button
-                data-cy="century"
-                className={cn('button mr-1', {
-                  'is-info': centuries.includes(century),
-                })}
-                key={century}
-                onClick={() => toggleCentury(century)}
-              >
-                {century}
-              </button>
-            ))}
-          </div>
-
-          <div className="level-right ml-4">
-            <button
-              data-cy="centuryALL"
-              className={cn('button is-success', {
-                'is-outlined': centuries.length,
-              })}
-              onClick={clearCenturies}
-            >
-              All
-            </button>
-          </div>
-        </div>
-      </div>
-
+      <SexFilter sex={sex} />
+      <QueryFilter query={query} onChange={handleQueryChange} />
+      <CenturyFilter
+        centuries={centuries}
+        centuryArr={centuryArr}
+        onToggleCentury={toggleCentury}
+        onClearCenturies={clearCenturies}
+      />
       <div className="panel-block">
         <SearchLink
           className="button is-link is-outlined is-fullwidth"
