+++ src/components/PeopleFilters/SexFilter/index.ts
@@ -1,0 +1,1 @@
+export * from './SexFilter';
